remove AppLayout support for multiple workspaces
